files renamed and components separated
