+++ package.json
@@ -6,7 +6,9 @@
   "scripts": {
     "dev": "vite",
     "build": "vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "storybook": "start-storybook -p 6006",
+    "build-storybook": "build-storybook"
   },
   "dependencies": {
     "@codemirror/lang-javascript": "^6.1.0",
@@ -31,6 +33,9 @@
     "@types/react": "^18.0.17",
     "@types/react-dom": "^18.0.6",
     "@vitejs/plugin-react": "^2.1.0",
-    "vite": "^3.1.0"
+    "vite": "^3.1.0",
+    "@storybook/react": "^6.5.12",
+    "@storybook/addon-actions": "^6.5.12",
+    "@storybook/addon-links": "^6.5.12"
   }
 }
