--- src/App.css
@@ -1,65 +1,0 @@
-body {
-	color: #666;
-	background: white !important;
-}
-
-.header {
-	background-color: white;
-  top: 0;
-  position: fixed;
-  width: 100%;
-  margin-bottom:2rem;
-}
-
-.header_logo {
-  padding-top: 10px;
-  padding-bottom: 5px;
-}
-
-.header_logo a {
-  color: #0074c1;
-  padding-left: 30px;
-  font-size: 1.2rem;
-}
-
-.header_logo a:hover {
-  color: #0074c1;
-}
-
-.footer{
-  font-size: 0.6rem;
-  position: fixed;
-  bottom: 0;
-  right: 10px;
-}
-.main-area{
-    margin-top:2.5rem;
-}
-
-.theme-selector{
-    position:fixed;
-    bottom:5;
-    left:10px;
-    
-}
-.split {
-  display: flex;
-  flex-direction: row;
-}
-
-.gutter {
-  background-color: #eee;
-  background-repeat: no-repeat;
-  background-position: 50%;
-}
-
-.gutter.gutter-horizontal {
-  cursor: col-resize;
-}
-
-iframe{
-  width: 100%;
-  height: 100%;
-  border: 0;
-
-}
