+++ src/components/app/app.css
@@ -1,0 +1,8 @@
+body {
+	color: #666;
+	background: white !important;
+}
+
+
+
+
