+++ src/components/code-editor/code-editor.css
@@ -1,0 +1,49 @@
+.main-area {
+  /* margin-top: 2.5rem; */
+}
+
+.theme-selector {
+  position: fixed;
+  bottom: 5;
+  left: 10px;
+
+}
+
+.split {
+  display: flex;
+  flex-direction: row;
+}
+
+.gutter {
+  background-color: #eee;
+  background-repeat: no-repeat;
+  background-position: 50%;
+}
+
+.gutter.gutter-horizontal {
+  cursor: col-resize;
+}
+
+iframe {
+  width: 100%;
+  height: 100%;
+  border: 0;
+
+}
+
+.cm-content,
+.cm-gutter {
+  min-height: 150px;
+}
+
+.cm-gutters {
+  margin: 1px;
+}
+
+.cm-scroller {
+  overflow: auto;
+}
+
+.cm-wrap {
+  border: 1px solid silver
+}
