+++ src/components/footer/footer.css
@@ -1,0 +1,6 @@
+.footer {
+  font-size: 0.6rem;
+  position: fixed;
+  bottom: 0;
+  right: 10px;
+}
