+++ src/components/footer/footer.jsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import './footer.css'
 
 const Footer = () => {
   return (
