+++ src/components/header/header.css
@@ -1,0 +1,22 @@
+.header {
+  background-color: white;
+  top: 0;
+  /* position: fixed; */
+  width: 100%;
+  /* margin-bottom: 2rem; */
+}
+
+.header_logo {
+  padding-top: 10px;
+  padding-bottom: 5px;
+}
+
+.header_logo a {
+  color: #0074c1;
+  padding-left: 30px;
+  font-size: 1.2rem;
+}
+
+.header_logo a:hover {
+  color: #0074c1;
+}
