+++ src/components/mainbody/mainbody.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import CodeEditor from './code-editor';
+import CodeEditor from '../code-editor/code-editor';
 
 const Mainbody = () => {
   return (
